environmentd: move HTTP SQL code into `sql` module
The top-level `http` module is only meant for code common to all HTTP
requests, like setting up the server and authenticating requests. The
code specific to the SQL endpoint belongs in the dedicated `sql` module.

This commit is pure code movement.
